Tugas4/task-service/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
mysql-connector-python==8.0.26
